+++ pkg/steps/stepexecution.go
@@ -211,7 +211,10 @@ func NewExecution(s Step, logger *logrus.Entry, fs afero.Fs, regionDeployType Re
 	}
 }
 
-func (s Step) InitExecution(logger *logrus.Entry, fs afero.Fs, regionDeployType RegionDeployType, region string, defaultStepOutputVariables map[string]map[string]string) (ExecutionConfig, error) {
+func (s Step) InitExecution(logger *logrus.Entry, fs afero.Fs,
+	regionDeployType RegionDeployType, region string,
+	defaultStepOutputVariables map[string]map[string]string) (
+	ExecutionConfig, error) {
 	exec := NewExecution(s, logger, fs, regionDeployType, region, defaultStepOutputVariables)
 
 	// set and create execution directory to enable safe concurrency
@@ -346,25 +349,54 @@ func (s Step) InitExecution(logger *logrus.Entry, fs afero.Fs, regionDeployType 
 
 	exec.TFBackend = GetBackendConfig(exec, ParseTFBackend)
 
-	HandleOverrides(exec.Logger, exec.Dir, exec.DeploymentRing)
+	// Handle any override files in the step and deployment ring. Deploy
+	// overrides will always copied over and they destroy overrides will be
+	// copied over only on Self-Destroy.
+	HandleDeployOverrides(exec.Logger, exec.Dir, exec.DeploymentRing)
+
+	if s.DeployConfig.SelfDestroy {
+		HandleDestroyOverrides(exec.Logger, exec.Dir, exec.DeploymentRing)
+	}
 
 	return exec, nil
 }
 
-// copy override configuration into execution working directory
-func HandleOverrides(logger *logrus.Entry, execDir string, deploymentRing string) {
-	ringOverrideFile := fmt.Sprintf("ring_%s_override.tf", strings.ToLower(deploymentRing))
-
-	src := filepath.Join(execDir, "override", ringOverrideFile)
-	dst := filepath.Join(execDir, ringOverrideFile)
+func handleOverride(logger *logrus.Entry, execDir string, fileName string) {
+	src := filepath.Join(execDir, "override", fileName)
+	dst := filepath.Join(execDir, fileName)
 
 	logger.Infof("Attempting to copy %s to %s", src, dst)
 
 	err := CopyFile(src, dst)
 
 	if err != nil && !os.IsNotExist(err) {
-		logger.WithError(err).Errorf("Overrides were not successfully set targeting %s", ringOverrideFile)
+		logger.WithError(err).Errorf(
+			"Overrides were not successfully set targeting %s", fileName)
 	}
+}
+
+// HandleDeployOverrides copy deploy override configurations into the
+// execution working directory
+func HandleDeployOverrides(logger *logrus.Entry, execDir string,
+	deploymentRing string) {
+	overrideFile := "override.tf"
+	ringOverrideFile := fmt.Sprintf("ring_%s_override.tf",
+		strings.ToLower(deploymentRing))
+
+	handleOverride(logger, execDir, overrideFile)
+	handleOverride(logger, execDir, ringOverrideFile)
+}
+
+// HandleDestroyOverrides copy destroy override configurations into the
+// execution working directory
+func HandleDestroyOverrides(logger *logrus.Entry, execDir string,
+	deploymentRing string) {
+	destroyOverrideFile := "destroy_override.tf"
+	destroyRingOverrideFile := fmt.Sprintf("destroy_ring_%s_override.tf",
+		strings.ToLower(deploymentRing))
+
+	handleOverride(logger, execDir, destroyOverrideFile)
+	handleOverride(logger, execDir, destroyRingOverrideFile)
 }
 
 // ExecuteStepDestroy destroys a step