+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Terrascale
 
+## 0.2.2 (July 27, 2020)
+
+### ENHANCEMENTS
+
+- Additional support for override files for deployments.
+  - `override.tf` files will be supported to all deployments, including Self-Destroy.
+  - `destroy_override.tf` files will be supported to Self-Destroy deployments.
+  - `destroy_ring_*ring-name*_override.tf` files will be added to Self-Destroy deployments for the specified deployment ring.
+
 ## 0.2.1 (July 21, 2020)
 
 ### BUG FIX